fix: rename video-web
