+++ packages/rax-player/src/index.js
@@ -1,2 +1,2 @@
-import H5Video from './video-h5';
+import H5Video from './video-web';
 export default H5Video;
